ai_agent_framework/requirements.txt
httpx>=0.25.0
PyYAML>=6.0
